Refactor to avoid duplicate code
